Implement MAC table update logic, and forwarding logic
Instead of having a fixed number of ports like a physical switch,
this virtual switch has a set of dynamic vports, one for each device
which has sent to the vswitch before. When a broadcast frame is received,
it is forwarded via every vport except the source vport.

If the destination MAC is not in the MAC table, and it is not
the broadcast MAC, the frame will be dropped as ARP resolution
is outside the scope of this project