Changed ioctl_write_ptr to take raw c_int pointer
linux/if_tun.h defines TUNSETIFF as follows:

define TUNSETIFF     _IOW('T', 202, int)

'T' means the operation relates to the tun/tap
driver, and 202 means interface flags are being set,
which was correct initially.

However I passed ifreq to the macro initially as that
is what i thought the call would take, however it taking
a pointer to an int instead changed the behaviour of the
ioctl syscall causing it to fail as there is a subtle
difference in the calling convention between the two.

This change means the syscall works as expected, and the
program works.