Added docker compose for platform independence
This repo contains tun/tap interface code which is
specific to Linux platforms. To allow this repo
to run on other platforms I have added a docker
compose file which uses rust alpine containers to
run the vswitch/vport code.

The docker compose also creates a shared network
for the containers to communicate over

The vport containers require the NET_ADMIN
permission to create and manage the tuntap
device